feat: use pip instead of apk
using pip instead of apk will allow any ansible module requirements to easily be added to the dockerfile. For example the pywinrm and other dependency's are only available via pip. It also allows the newest version available. So it makes sense for all python depdencies to come from the same source.